создал ВМ для тестового робота и перекинул методы
