+++ project/OsEngine/OsEngine.csproj
@@ -837,6 +837,7 @@
     <Compile Include="OsaExtension\MVVM\ViewModels\ChangeEmitentVM.cs" />
     <Compile Include="OsaExtension\MVVM\ViewModels\IRobotVM.cs" />
     <Compile Include="OsaExtension\MVVM\ViewModels\MainWindowRobWpfVM.cs" />
+    <Compile Include="OsaExtension\MVVM\ViewModels\TestRobVM.cs" />
     <Compile Include="OsaExtension\MVVM\View\ChengeEmitendWidow.xaml.cs">
       <DependentUpon>ChengeEmitendWidow.xaml</DependentUpon>
     </Compile>
