+++ project/OsEngine/OsaExtension/MVVM/ViewModels/MainWindowRobWpfVM.cs
@@ -8,6 +8,7 @@ using OsEngine.OsTrader.Panels;
 using OsEngine.Robots.Trend;
 using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Diagnostics;
 using System.Linq;
 using System.Text;
@@ -22,8 +23,7 @@ namespace OsEngine.OsaExtension.MVVM.ViewModels
     {
         public MainWindowRobWpfVM() 
         {
-            ServerMaster.ServerCreateEvent += ServerMaster_ServerCreateEvent;
-            
+                        
         }
 
         /// <summary>
@@ -32,70 +32,11 @@ namespace OsEngine.OsaExtension.MVVM.ViewModels
         public static ChengeEmitendWidow ChengeEmitendWidow = null;
 
         #region Property ==================================================================
-
-        /// <summary>
-        /// Сервер 
-        /// </summary>
-        public IServer Server
-        {
-            get => _server;
-            set
-            {
-                if (Server != null)
-                {
-                    UnSubscribeToServer();
-                    _server = null;
-                }
-                _server = value;
-                OnPropertyChanged(nameof(ServerType));
-
-                SubscribeToServer(); // подключаемя к бир
-            }
-        }
-        private IServer _server = null;
-
-        /// <summary>
-        /// Рыночная цена бумаги 
-        /// </summary>
-        public decimal Price
-        {
-            get => _price;
-            set
-            {
-                _price = value;
-                OnPropertyChanged(nameof(Price));
-            }
-        }
-        private decimal _price;
+        public ObservableCollection<EnvelopTrend> Robots { get; set; } = new ObservableCollection<EnvelopTrend>();
 
         #endregion
-        /// <summary>
-        /// робот
-        /// </summary>
-        public EnvelopTrend Bot { get; set; }
-
-        /// <summary>
-        /// робот
-        /// </summary>
-        public string NamSecuriti { get; set; }
-
 
         #region Commands ===========================================================
-
-
-        private DelegateCommand _commandSelectSecurity;
-        public DelegateCommand CommandSelectSecurity
-        {
-            get
-            {
-                if (_commandSelectSecurity == null)
-                {
-                    _commandSelectSecurity = new DelegateCommand(SelectSecurity);
-                }
-                return _commandSelectSecurity;
-            }
-        }
-        #endregion
 
 
         //public event GridRobotVM.selectedSecurity OnSelectedSecurity;
@@ -129,75 +70,7 @@ namespace OsEngine.OsaExtension.MVVM.ViewModels
                 return commandСreateBot;
             }
         }
-
-        #region Metods ============================================================
-
-        /// <summary>
-        /// выбрать бумагу
-        /// </summary>
-        void SelectSecurity(object o)
-        {
-            if (MainWindowRobWpfVM.ChengeEmitendWidow != null)
-            {
-                return;
-            }
-            MainWindowRobWpfVM.ChengeEmitendWidow = new ChengeEmitendWidow();
-            MainWindowRobWpfVM.ChengeEmitendWidow.ShowDialog();
-            MainWindowRobWpfVM.ChengeEmitendWidow = null;
-            if (_server != null)
-            {
-                if (_server.ServerType == ServerType.Binance
-                    || _server.ServerType == ServerType.BinanceFutures)
-                {
-                    // IsChekCurrency = true;
-                }
-                //else IsChekCurrency = false;
-            }
-        }
-
-        /// <summary>
-        ///  подключиться к серверу
-        /// </summary>
-        private void SubscribeToServer()
-        {
-            //_server.NewMyTradeEvent += Server_NewMyTradeEvent;
-            //_server.NewOrderIncomeEvent += Server_NewOrderIncomeEvent;
-            //_server.NewCandleIncomeEvent += Server_NewCandleIncomeEvent;
-            _server.NewTradeEvent += _server_NewTradeEvent;
-            //_server.SecuritiesChangeEvent += _server_SecuritiesChangeEvent;
-            //_server.PortfoliosChangeEvent += _server_PortfoliosChangeEvent;
-            //_server.NewBidAscIncomeEvent += _server_NewBidAscIncomeEvent;
-            //_server.ConnectStatusChangeEvent += _server_ConnectStatusChangeEvent;            
-        }
-
-        /// <summary>
-        ///  отключиться от сервера 
-        /// </summary>
-        private void UnSubscribeToServer()
-        {
-            //_server.NewMyTradeEvent -= Server_NewMyTradeEvent;
-            //_server.NewOrderIncomeEvent -= Server_NewOrderIncomeEvent;
-            //_server.NewCandleIncomeEvent -= Server_NewCandleIncomeEvent;
-            _server.NewTradeEvent -= _server_NewTradeEvent;
-            //_server.SecuritiesChangeEvent -= _server_SecuritiesChangeEvent;
-            //_server.PortfoliosChangeEvent -= _server_PortfoliosChangeEvent;
-            //_server.NewBidAscIncomeEvent -= _server_NewBidAscIncomeEvent;
-            //_server.ConnectStatusChangeEvent -= _server_ConnectStatusChangeEvent;
-        }
-
-        private void ServerMaster_ServerCreateEvent(IServer server)
-        {
-            if (_server == null) return;
-
-            _server.NewTradeEvent += _server_NewTradeEvent;
-        }
-
-        private void _server_NewTradeEvent(List<Trade> trades)
-        {
-            Trade trade = trades.Last();
-
-            Price = trade.Price;
-        }
+        #endregion
 
         /// <summary>
         ///  подключение к серверу 
@@ -217,9 +90,10 @@ namespace OsEngine.OsaExtension.MVVM.ViewModels
 
             EnvelopTrend bot = new EnvelopTrend(name, StartProgram.IsOsTrader);
 
-            Bot = bot;
+            Robots.Add(bot);
+            //Bot = bot;
 
-            NamSecuriti = bot.TabsSimple[0].Securiti.Name;
+            //NamSecuriti = bot.TabsSimple[0].Securiti.Name;
 
            
         }
@@ -227,8 +101,5 @@ namespace OsEngine.OsaExtension.MVVM.ViewModels
         public event selectedSecurity OnSelectedSecurity;
 
 
-
-
-        #endregion
     }
 }
